requirements.txt
fitz==0.0.1.dev2
Pillow==11.2.1
